# docs/openapi/velontri_openapi.yaml
openapi: 3.1.0
info:
  title: Velontri API
  version: 1.0.0
  description: |
    Complete API for all 14 Velontri microservices.
    All responses use a standard envelope: { success, message, data, meta }
    All errors use: { success: false, error: { code, message, field }, request_id }
  contact:
    name: Velontri Engineering
    email: dev@velontri.com
servers:
  - url: https://api.velontri.com/api/v1
    description: Production
  - url: https://api.staging.velontri.com/api/v1
    description: Staging
  - url: http://localhost:8001/api/v1
    description: Auth service (local)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT. Access token TTL is 15 minutes.
...
paths:
...
  /search:
    get:
      tags: [Search]
      summary: Keyword + filter search
      parameters:
        - { name: q,          in: query, required: true, schema: { type: string }, example: "Toyota Camry Lagos" }
        - { name: page,       in: query, schema: { type: integer, default: 1 } }
        - { name: page_size,  in: query, schema: { type: integer, default: 20 } }
        - { name: price_min,  in: query, schema: { type: number } }
        - { name: price_max,  in: query, schema: { type: number }, example: 10000000 }
        - { name: category,   in: query, schema: { type: string }, example: "vehicles" }
        - { name: city,       in: query, schema: { type: string }, example: "Lagos" }
        - { name: country,    in: query, schema: { type: string }, example: "NG" }
        - { name: lat,        in: query, schema: { type: number }, example: 6.5244 }
        - { name: lon,        in: query, schema: { type: number }, example: 3.3792 }
        - { name: radius_km,  in: query, schema: { type: number }, example: 10 }
        - { name: condition,  in: query, schema: { type: string, enum: [new, used, refurbished] } }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              example:
                data:
                  hits: [{ id: "uuid", title: "Toyota Camry 2019", score: 0.95 }]
                  total: 12
                  page: 1
  /search/autocomplete:
    get:
      tags: [Search]
      summary: Autocomplete suggestions
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 }, example: "Toyo" }
      responses:
        '200':
          description: Suggestions
          content:
            application/json:
              example:
                data: { suggestions: ["Toyota Camry", "Toyota Corolla", "Toyota Hilux"] }
  /search/voice:
    post:
      tags: [Search]
      summary: Voice search via audio upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio: { type: string, format: binary, description: "Audio file (wav, mp3, ogg, webm)" }
      responses:
        '200': { description: Transcribed and searched }
  /search/ai:
    post:
      tags: [Search]
      summary: AI natural-language search
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:     { type: string, example: "I need a 3-bedroom flat in Ikeja under 500k per year" }
                page:      { type: integer, default: 1 }
                page_size: { type: integer, default: 20 }
      responses:
        '200': { description: AI search results }

  # ── WALLET ───────────────────────────────────────────────────────────────────
...